finished project logic

+++ src/main/java/app/taskLogic/ProjectController.java
@@ -5,8 +5,11 @@ package app.taskLogic;
 
 import app.auth.security.jwt.JwtUtils;
 import app.taskLogic.models.Project;
+import app.taskLogic.models.Task;
 import app.taskLogic.request.CreateProjectRequest;
 import app.taskLogic.service.ProjectService;
+import app.taskLogic.service.TaskService;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,33 +17,51 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @RestController
+@SecurityRequirement(name = "JWT")
 @RequestMapping("/projects")
+@PreAuthorize("hasRole('ROLE_ADMIN')")
+@CrossOrigin(origins = "*")
 public class ProjectController {
     @Autowired
     private ProjectService projectService;
 
     @Autowired
+    TaskService taskService;
+    @Autowired
     private JwtUtils jwtUtils;
 
-    // Создание нового проекта
     @PostMapping("/create")
-    public ResponseEntity<Project> createProject(@RequestBody Project project) {
 
-        Project createdProject = projectService.createProject(project);
-        return new ResponseEntity<>(createdProject, HttpStatus.CREATED);
+    public ResponseEntity<Project> createProject(@RequestBody CreateProjectRequest projectRequest) {
+        try {
+            Project project = new Project();
+            project.setName(projectRequest.getName());
+            Project createdProject = projectService.createProject(project);
+            return new ResponseEntity<>(createdProject, HttpStatus.CREATED);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
-
 
     @DeleteMapping("/{projectId}")
-    @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<Void> deleteProject(@PathVariable Long projectId,
-    @RequestHeader(value = "Authorization") String authorizationHeader
+    public ResponseEntity<String> deleteProject(@PathVariable Long projectId) {
+        try {
+            Project project = projectService.getProjectById(projectId).orElse(null);
+            if (project == null) {
+                return new ResponseEntity<>("Проект не найден", HttpStatus.NOT_FOUND);
+            }
 
-    ) {
-        // Возможно, вам нужно провести аутентификацию пользователя и проверить его права доступа
-        // В этом примере, просто передаем идентификатор проекта сервису для удаления
-        projectService.deleteProject(projectId);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            List<Task> fds = new ArrayList<Task>() ;
+
+            projectService.deleteProject(projectId);
+            return new ResponseEntity<>("Проект удален", HttpStatus.NO_CONTENT);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
+
 }
